@@ -1,5 +1,5 @@
 module github.com/icholy/digest
 
-go 1.14
+go 1.16
 
 require gotest.tools/v3 v3.0.2
